+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Custom About window with version info and links (Website, GitHub, Documentation)
+
 ## [0.9.2] - 2026-02-28
 
 ### Fixed